refactor: followup cleanup for #65758 fallout
Fixes some issues left behind by recent refactor
