+++ packages/forms/signals/src/api/rules/validation/index.ts
@@ -14,5 +14,8 @@ export * from './min_length';
 export * from './pattern';
 export * from './required';
 export * from './standard_schema';
+export * from './validate';
 export * from './validate_async';
 export * from './validate_http';
+export * from './validate_tree';
+export * from './validation_errors';
